woop

+++ src/runtime/io.cpp
@@ -1540,7 +1540,7 @@ extern "C" LEAN_EXPORT obj_res lean_io_exit(uint8_t code, obj_arg /* w */) {
 }
 
 extern "C" LEAN_EXPORT obj_res lean_io_quick_exit(uint8_t code, obj_arg /* w */) {
-    std::quick_exit(code);
+    std::quick_exit((int)code);
 }
 
 extern "C" LEAN_EXPORT obj_res lean_runtime_mark_multi_threaded(obj_arg a, obj_arg /* w */) {
